Follow zenux-services' changes on DSP variables
There have been so many random looking test results (and let's not think about
target): Now we should have solid tests on DSP variables with the code we got.

+++ modules-common/service-interfaces/testlib/testdspinterface.h
@@ -25,7 +25,7 @@ private:
     struct TVarsWritten {
         int transcationCount;
         QString groupVarName;
-        float dataWritten;
+        double dataWritten; // enough precision for uint (32 bit)
     };
     QList<TVarsWritten> m_valuesWritten;
     int m_transactionCount = 0;
